Move single thread tests to JVM Suites
