+++ monix-eval/shared/src/test/scala/monix/eval/TaskLocalSuite.scala
@@ -268,55 +268,4 @@ object TaskLocalSuite extends SimpleTestSuite {
 
     t.runToFutureOpt
   }
-
-  testAsync("TaskLocal.isolate should properly isolate during async boundaries") {
-    implicit val s = TracingScheduler(Scheduler.singleThread("local-test"))
-      .withExecutionModel(ExecutionModel.AlwaysAsyncExecution)
-
-    val local = Local(0)
-    val test = for {
-      _ <- Task.evalAsync(local := 50)
-      _ <- TaskLocal.isolate {
-        Task.evalAsync(local := 100)
-      }
-      v <- Task.evalAsync(local())
-      _ <- Task.now(assertEquals(v, 50))
-    } yield ()
-
-    test.runToFutureOpt
-  }
-
-  testAsync("TaskLocal.isolate should properly isolate during async boundaries on error") {
-    implicit val s = TracingScheduler(Scheduler.singleThread("local-test"))
-      .withExecutionModel(ExecutionModel.AlwaysAsyncExecution)
-
-    val local = Local(0)
-    val test = for {
-      _ <- Task.evalAsync(local := 50)
-      _ <- TaskLocal.isolate {
-        Task.evalAsync(local := 100).flatMap(_ => Task.raiseError(DummyException("boom")))
-      }.attempt
-      v <- Task.evalAsync(local())
-      _ <- Task.now(assertEquals(v, 50))
-    } yield ()
-
-    test.runToFutureOpt
-  }
-
-  testAsync("TaskLocal.isolate should properly isolate during async boundaries on cancelation") {
-    implicit val s = TracingScheduler(Scheduler.singleThread("local-test"))
-      .withExecutionModel(ExecutionModel.AlwaysAsyncExecution)
-
-    val local = Local(0)
-    val test = for {
-      _ <- Task.evalAsync(local := 50)
-      _ <- TaskLocal.isolate {
-        Task.evalAsync(local := 100).start.flatMap(_.cancel)
-      }
-      v <- Task.evalAsync(local())
-      _ <- Task.now(assertEquals(v, 50))
-    } yield ()
-
-    test.runToFutureOpt
-  }
 }
